Using the controller in order to view the articles
ArticleController is able to show one or the list of the articles. Index.php is able to choose the good function of the ArticleController and display the list of articles if there is an error. And to finish the function fucntion showAction is able to display one article and is able to display an error message if the url is not valid